requirements.txt
Django==3.1
djangorestframework
djongo==1.3.6
dnspython==2.5.0
factory-boy
Faker
pymongo==3.12.0
sqlparse==0.2.4
python-dotenv
pytest
pytest-django